Modul_4/Cinescope/requirements.txt
pytest
requests
faker
pydantic"
jsonschema
pytest-mock
sqlalchemy 
psycopg2
allure-pytest
allure-python-commons
pytest-check
pytest-rerunfailures
playwright